pyproject.toml
[project]
name = "nomad-stac-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "geopandas[all]>=1.1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.1",
    "pystac[jinja2,orjson,urllib3,validation]>=1.14.0",
    "pyyaml>=6.0.2",
    "stac-validator>=3.10.1",
    "tqdm>=4.67.1",
    "typer>=0.17.4",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.0",
]
